[Modify] 토론방 utils 및 types 이동
