+++ utils/debates/debateroom/toggle.ts
@@ -1,4 +1,4 @@
-import { IDebateroom } from "../types";
+import { IDebateroom } from "../../../types";
 
 export const toggleMic = ({
   stream,
